Use onCreateView itself to find and hold views in members
‘Error inflating fragment’ is very general and flexible. This it time
it caused the details screen to crash before loading. It was caused by
using getView() instead of the just inflated out rootView in
onCreateView to findById.

DetailFragment in a separate file.

Tiny adjustments in layout.
