+++ app/src/main/res/layout/activity_detail.xml
@@ -1,7 +1,7 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/container"
-    android:name="com.example.pratu16x7.sunshinefromscratch.DetailActivity$DetailFragment"
+    class="com.example.pratu16x7.sunshinefromscratch.DetailFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
